more consolidation of MaterialX nodes.
