+++ src/nodes/materialx/MaterialXColor.js
@@ -1,9 +1,6 @@
-// Three.js Transpiler
-// https://github.com/AcademySoftwareFoundation/MaterialX/blob/main/libraries/stdlib/genglsl/lib/mx_hsv.glsl
-
-import { int, float, vec3, If, Fn } from '../../tsl/TSLBase.js';
-import { add } from '../../math/OperatorNode.js';
-import { floor, trunc, max, min } from '../../math/MathNode.js';
+import { int, float, vec3, If, Fn } from '../tsl/TSLBase.js';
+import { add } from '../math/OperatorNode.js';
+import { floor, trunc, max, min } from '../math/MathNode.js';
 
 export const mx_hsvtorgb = /*@__PURE__*/ Fn( ( [ hsv ] ) => {
