+++ src/nodes/materialx/MaterialXColorTransform.js
@@ -1,9 +1,6 @@
-// Three.js Transpiler
-// https://github.com/AcademySoftwareFoundation/MaterialX/blob/main/libraries/stdlib/genglsl/lib/mx_transform_color.glsl
-
-import { bvec3, vec3, Fn } from '../../tsl/TSLBase.js';
-import { greaterThan } from '../../math/OperatorNode.js';
-import { max, pow, mix } from '../../math/MathNode.js';
+import { bvec3, vec3, Fn } from '../tsl/TSLBase.js';
+import { greaterThan } from '../math/OperatorNode.js';
+import { max, pow, mix } from '../math/MathNode.js';
 
 export const mx_srgb_texture_to_lin_rec709 = /*@__PURE__*/ Fn( ( [ color_immutable ] ) => {
